+++ .github/workflows/release.yaml
@@ -26,7 +26,7 @@ jobs:
 
       - name: Determine version bump
         run: |
-          ./bump-version.sh
+          ./scripts/bump-version.sh
 
       - name: Check if version bump should be ignored
         id: bump-check
