+++ internal/registry/testdata/route_config.yaml
