+++ scripts/bump-version.sh
